backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiohttp==3.9.1
python-nmap==0.7.1
httpx==0.25.2
python-multipart==0.0.6
loguru==0.7.2
pyyaml==6.0.1

